Use OS and fallback DNS servers in DNSSEC resolver (#1736)
* DNSSEC refactoring
