+++ cmd/catchupsrv/download.go
@@ -194,7 +194,7 @@ func download() {
 			}
 
 			var resolver tools_network.Resolver
-			resolver.DNSAddress = *dnsAddr
+			resolver.SetFallbackResolverAddress(*dnsAddr)
 			_, records, err = resolver.LookupSRV(context.Background(), "algobootstrap", "tcp", bootstrapID)
 			if err != nil {
 				panic(err)
