+++ network/dialer.go
@@ -37,7 +37,7 @@ type Dialer struct {
 
 // makeRateLimitingDialer creates a rate limiting dialer that would limit the connections
 // according to the entries in the phonebook.
-func makeRateLimitingDialer(phonebook Phonebook, resolver *dnssec.Resolver) Dialer {
+func makeRateLimitingDialer(phonebook Phonebook, resolver dnssec.ResolverIf) Dialer {
 	var innerDialer netDialer = &net.Dialer{
 		Timeout:   30 * time.Second,
 		KeepAlive: 30 * time.Second,
