+++ tools/network/dnssec/resolver.go
@@ -14,6 +14,27 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with go-algorand.  If not, see <https://www.gnu.org/licenses/>.
 
+/*
+Package dnssec provides net.Resolver-compatible methods for DNSSEC.
+
+Fully DNSSEC-compliant:
+LookupIPAddr, LookupCNAME, LookupSRV, LookupTXT, LookupMX, LookupNS, LookupTLSA
+
+Fallbacks to net.DefaultResolver:
+LookupAddr, LookupPort, LookupHost
+
+The package uses miekg/dns for low-level DNS intractions and DNS messages parsing.
+
+References
+1. DNS https://tools.ietf.org/html/rfc1035
+2. DNS clarifications https://tools.ietf.org/html/rfc2181
+3. DNSSEC proto change https://tools.ietf.org/html/rfc4035
+4. DNSSEC RR change https://tools.ietf.org/html/rfc4034
+5. DNSSEC clarifications https://tools.ietf.org/html/rfc6840
+6. DNSSEC keys management https://tools.ietf.org/html/rfc6781
+7. DNS SRV https://tools.ietf.org/html/rfc2782
+
+*/
 package dnssec
 
 import (
@@ -23,53 +44,62 @@ import (
 	"time"
 
 	"github.com/miekg/dns"
+
+	"github.com/algorand/go-algorand/logging"
 )
 
-const defaultMaxHops = 10
-
-// List of DNSSEC-aware servers
-// CloudFlare: 1.1.1.1:53 1.0.0.1:53
-// Google: 8.8.8.8:53 8.8.4.4:53
-// Yandex 77.88.8.8:53 77.88.8.1:53
-// Comodo 8.26.56.26:53 8.20.247.20:53
-//
-// Other - no DNSSEC
-// OpenDNS 208.67.222.222:53
-// Baidu 180.76.76.76:53
-// Alibaba 223.6.6.6:53
-
-var defaultDnssecAwareNSServers = []string{"1.1.1.1:53", "8.8.8.8:53", "77.88.8.8:53", "8.26.56.26:53"}
+// ResolverIf represents net.Resolver-compatible interface
+type ResolverIf interface {
+	LookupAddr(ctx context.Context, addr string) (names []string, err error)
+	LookupCNAME(ctx context.Context, host string) (cname string, err error)
+	LookupHost(ctx context.Context, host string) (addrs []string, err error)
+	LookupIPAddr(ctx context.Context, host string) ([]net.IPAddr, error)
+	LookupMX(ctx context.Context, name string) ([]*net.MX, error)
+	LookupNS(ctx context.Context, name string) ([]*net.NS, error)
+	LookupPort(ctx context.Context, network, service string) (port int, err error)
+	LookupSRV(ctx context.Context, service, proto, name string) (cname string, addrs []*net.SRV, err error)
+	LookupTXT(ctx context.Context, name string) ([]string, error)
+}
 
 // Resolver provides DNSSEC resolution
 type Resolver struct {
-	resolver   netResolverIf
+	client     Querier
 	trustChain *trustChain
 	maxHops    int
 }
 
-// DefaultResolver with one DNS server and 1 second timeout
-var DefaultResolver = MakeDnssecResolver(defaultDnssecAwareNSServers, time.Second)
-
 // MakeDnssecResolver return resolver from given NS servers and timeout duration
-func MakeDnssecResolver(servers []string, timeout time.Duration) (r Resolver) {
-	rs := &resolverImpl{readTimeout: timeout, servers: servers, rootAnchor: rootAnchorXML}
-
-	if len(rs.servers) == 0 {
-		rs.servers = append(rs.servers, defaultDnssecAwareNSServers...)
-	}
-	r.resolver = rs
-	r.trustChain = makeTrustChain(r.resolver)
-	r.maxHops = defaultMaxHops
-	return
+func MakeDnssecResolver(servers []ResolverAddress, timeout time.Duration) ResolverIf {
+	dc := MakeDNSClient(servers, timeout)
+	tc := &QueryWrapper{dc}
+	return &Resolver{client: dc, trustChain: makeTrustChain(tc), maxHops: DefaultMaxHops}
 }
 
-type netResolverIf interface {
-	query(ctx context.Context, domain string, qtype uint16) (resp *dns.Msg, err error)
-	queryRRSet(ctx context.Context, domain string, qtype uint16) ([]dns.RR, []dns.RRSIG, error)
-	rootTrustAnchor() ([]dns.DS, error)
+// MakeDefaultDnssecResolver returns a resolver with all possible DNS servers:
+// system, fallback, default
+func MakeDefaultDnssecResolver(fallbackAddress string, log logging.Logger) ResolverIf {
+	servers, timeout, err := SystemConfig()
+	if err != nil {
+		log.Debugf("retrieving system config failed with %s", err.Error())
+		timeout = DefaultTimeout
+	}
 
-	// test functions
-	serverList() []string
+	var fallback string
+	var dnsIPAddr *net.IPAddr
+	if dnsIPAddr, err = net.ResolveIPAddr("ip", fallbackAddress); err != nil {
+		log.Debugf("resolving fallback %s failed with %s", fallbackAddress, err.Error())
+	} else {
+		fallback = dnsIPAddr.String()
+	}
+
+	servers = append(servers, MakeResolverAddress(fallback, "53"))
+	servers = append(servers, DefaultDnssecAwareNSServers...)
+	return MakeDnssecResolver(servers, timeout)
+}
+
+// EffectiveResolverDNS return list of active DNS servers
+func (r *Resolver) EffectiveResolverDNS() (servers []ResolverAddress) {
+	return r.client.(*dnsClient).servers
 }
 
 // TLSARec represents TLSA record content
@@ -80,22 +110,22 @@ type TLSARec struct {
 	Certificate  string `dns:"hex"`
 }
 
-// lookupImpl makes DNS request for a zone and verifies response signature
-func (r *Resolver) lookupImpl(ctx context.Context, name string, qt uint16) (rrSet []dns.RR, err error) {
-	rrSet, rrSig, err := r.resolver.queryRRSet(ctx, name, qt)
+// lookup makes DNS request for a zone and verifies response signature
+func (r *Resolver) lookup(ctx context.Context, name string, qt uint16) (rrSet []dns.RR, err error) {
+	rrSet, rrSig, err := r.client.QueryRRSet(ctx, name, qt)
 	if err != nil {
 		return
 	}
-	err = r.trustChain.authenticate(ctx, rrSet, rrSig)
+	err = r.trustChain.Authenticate(ctx, rrSet, rrSig)
 	return
 }
 
-// lookupImplCnameAware like lookupImpl makes DNS request for a zone and verifies response signature
+// lookupCnameAware requests a zone and verifies response signature similarly to lookup,
 // but also it is aware about possible A/CNAME entries for A-requests and supposed to be used for CNAME alias resolution.
 // if CNAME signature presents for requested domain name, then consider the response as CNAME
 // and extract CNAME record(s) and its RRSIG
-func (r *Resolver) lookupImplCnameAware(ctx context.Context, name string, qt uint16) (rrSet []dns.RR, err error) {
-	rrSet, rrSig, err := r.resolver.queryRRSet(ctx, name, qt)
+func (r *Resolver) lookupCnameAware(ctx context.Context, name string, qt uint16) (rrSet []dns.RR, err error) {
+	rrSet, rrSig, err := r.client.QueryRRSet(ctx, name, qt)
 	if err != nil {
 		return
 	}
@@ -104,7 +134,7 @@ func (r *Resolver) lookupImplCnameAware(ctx context.Context, name string, qt uin
 	// NS can return CNAME for A request if there are no A but CNAME
 	// in real world such A-response can also have second-level CNAME alias and A records with signatures
 	//
-	// so that leave only CNAME RR matching to requested name to verify signature and return the filtered set
+	// so that we leave only CNAME RR matching to requested name to verify signature and return the filtered set
 	if qt == dns.TypeA || qt == dns.TypeAAAA {
 		newRRSig := make([]dns.RRSIG, 0, len(rrSig))
 		for _, sig := range rrSig {
@@ -136,12 +166,12 @@ func (r *Resolver) lookupImplCnameAware(ctx context.Context, name string, qt uin
 			rrSig = newRRSig
 		}
 	}
-	err = r.trustChain.authenticate(ctx, rrSet, rrSig)
+	err = r.trustChain.Authenticate(ctx, rrSet, rrSig)
 	return
 }
 
 // LookupIPAddr resolves a given hostname to ipv4 or ipv6 address following CNAME aliaces
-func (r *Resolver) lookupIPAddrImpl(ctx context.Context, hostname string) (cname string, addrs []net.IPAddr, err error) {
+func (r *Resolver) lookupIPAddr(ctx context.Context, hostname string) (cname string, addrs []net.IPAddr, err error) {
 	var rrSet []dns.RR
 	nextName := hostname
 	seen := make(map[string]bool)
@@ -151,9 +181,9 @@ func (r *Resolver) lookupIPAddrImpl(ctx context.Context, hostname string) (cname
 			return
 		}
 
-		if rrSet, err = r.lookupImplCnameAware(ctx, nextName, dns.TypeA); err != nil {
+		if rrSet, err = r.lookupCnameAware(ctx, nextName, dns.TypeA); err != nil {
 			var err2 error
-			if rrSet, err2 = r.lookupImplCnameAware(ctx, nextName, dns.TypeAAAA); err2 != nil {
+			if rrSet, err2 = r.lookupCnameAware(ctx, nextName, dns.TypeAAAA); err2 != nil {
 				return // return original error
 			}
 		}
@@ -180,13 +210,13 @@ func (r *Resolver) lookupIPAddrImpl(ctx context.Context, hostname string) (cname
 
 // LookupIPAddr resolves a given hostname to ipv4 or ipv6 address
 func (r *Resolver) LookupIPAddr(ctx context.Context, host string) (addrs []net.IPAddr, err error) {
-	_, addrs, err = r.lookupIPAddrImpl(ctx, host)
+	_, addrs, err = r.lookupIPAddr(ctx, host)
 	return
 }
 
 // LookupCNAME returns CNAME record content for a given name
 func (r *Resolver) LookupCNAME(ctx context.Context, host string) (cname string, err error) {
-	cname, _, err = r.lookupIPAddrImpl(ctx, host)
+	cname, _, err = r.lookupIPAddr(ctx, host)
 	return
 }
 
@@ -201,7 +231,7 @@ func (r *Resolver) LookupSRV(ctx context.Context, service, proto, name string) (
 	}
 
 	var rrSet []dns.RR
-	if rrSet, err = r.lookupImpl(ctx, fullName, dns.TypeSRV); err != nil {
+	if rrSet, err = r.lookup(ctx, fullName, dns.TypeSRV); err != nil {
 		return
 	}
 
@@ -231,7 +261,7 @@ func (r *Resolver) LookupSRV(ctx context.Context, service, proto, name string) (
 // LookupMX returns MX records content for a given name
 func (r *Resolver) LookupMX(ctx context.Context, name string) (addrs []*net.MX, err error) {
 	var rrSet []dns.RR
-	if rrSet, err = r.lookupImpl(ctx, name, dns.TypeMX); err != nil {
+	if rrSet, err = r.lookup(ctx, name, dns.TypeMX); err != nil {
 		return
 	}
 
@@ -253,7 +283,7 @@ func (r *Resolver) LookupMX(ctx context.Context, name string) (addrs []*net.MX, 
 // LookupNS returns NS records content for a given name
 func (r *Resolver) LookupNS(ctx context.Context, name string) (addrs []*net.NS, err error) {
 	var rrSet []dns.RR
-	if rrSet, err = r.lookupImpl(ctx, name, dns.TypeNS); err != nil {
+	if rrSet, err = r.lookup(ctx, name, dns.TypeNS); err != nil {
 		return
 	}
 
@@ -274,7 +304,7 @@ func (r *Resolver) LookupNS(ctx context.Context, name string) (addrs []*net.NS, 
 // LookupTXT returns TXT records content for a given name
 func (r *Resolver) LookupTXT(ctx context.Context, name string) (addrs []string, err error) {
 	var rrSet []dns.RR
-	if rrSet, err = r.lookupImpl(ctx, name, dns.TypeTXT); err != nil {
+	if rrSet, err = r.lookup(ctx, name, dns.TypeTXT); err != nil {
 		return
 	}
 
@@ -300,7 +330,7 @@ func (r *Resolver) LookupTLSA(ctx context.Context, service, proto, name string) 
 	}
 
 	var rrSet []dns.RR
-	if rrSet, err = r.lookupImpl(ctx, fullName, dns.TypeTLSA); err != nil {
+	if rrSet, err = r.lookup(ctx, fullName, dns.TypeTLSA); err != nil {
 		return
 	}
 
@@ -334,9 +364,4 @@ func (r *Resolver) LookupAddr(ctx context.Context, addr string) (names []string,
 // LookupHost looks up the given host using the local resolver. It returns a slice of that host's addresses.
 func (r *Resolver) LookupHost(ctx context.Context, host string) (addrs []string, err error) {
 	return net.DefaultResolver.LookupHost(ctx, host)
-}
-
-// LookupSRV is convenience function using default dnssec resolver
-func LookupSRV(service, proto, name string) (cname string, addrs []*net.SRV, err error) {
-	return DefaultResolver.LookupSRV(context.Background(), service, proto, name)
 }
